Use additional power of two parameter.
The optimal, work-efficient algorithm assumes that the buffer has size
n, where n is a power of two.  If n happens not to be a power of two,
we "pad" the work array to fill it to the nearest larger power of two.
The additional number of elements used to pad is still linear.