Use additional temporary buffer for mapped data + cosmetics.
Fixes a bug where we weren't actually using the result of mapping 1s
and 0s to the input, and just reusing the raw input.  Now we create a
new temporary buffer on the GPU to hold it.

When used in the operators, this additional GPU buffer could be
allcoated a priori and reused, instead of being allocated every time.