Correct fundamental flaws in emitter algorithm.
By sending nullptr, we accitentally sent the value EOS, which
coincidentially has the same value.  This incorrectly signaled the end
of the stream, which isn't actually the case.  Also, the input should
be deleted only in case of a GPU buffer (the operator has input coming
directly from the GPU), it shouldn't be deleted if it's a single
tuple (it is re-sent right away to the assigned worker).