Conditionally define gpuErrChk macro.
Following the same idea behind the standard C assert macro, we only
define gpuErrChk to actually check for errors if the NDEBUG macro is
not defined.  If it is, then gpuErrChk simply expands to the argument
itself (wrapped in parenthesis, because with macros, you never know).
Unlike the standard C assert macro, which expands to the empty
statement if NDEBUG is defined, this makes it possible to use
gpuErrChk with expressions involving side effects and still have them
work as intended when not adding debug code.